Refactor: Move Destination client code into its own module
