+++ src/control/destination/background/mod.rs
@@ -7,17 +7,13 @@ use std::iter::IntoIterator;
 use std::net::SocketAddr;
 use std::time::{Instant, Duration};
 
-use bytes::Bytes;
 use futures::{
     future,
     sync::mpsc,
     Async, Future, Poll, Stream,
 };
-use h2;
-use http;
 use tower_grpc as grpc;
-use tower_h2::{self, BoxBody, HttpService, RecvBody};
-use tower_reconnect::Reconnect;
+use tower_h2::{BoxBody, HttpService, RecvBody};
 
 use linkerd2_proxy_api::destination::client::Destination;
 use linkerd2_proxy_api::destination::update::Update as PbUpdate2;
@@ -34,33 +30,20 @@ use control::{
     cache::{Cache, CacheChange, Exists},
     fully_qualified_authority::FullyQualifiedAuthority,
     remote_stream::{Receiver, Remote},
-    util::{AddOrigin, Backoff, LogErrors},
 };
 use dns::{self, IpAddrListFuture};
 use telemetry::metrics::DstLabels;
-use transport::{DnsNameAndPort, HostAndPort, LookupAddressAndConnect};
-use timeout::Timeout;
+use transport::{DnsNameAndPort, HostAndPort};
 use transport::tls;
 use conditional::Conditional;
-use watch_service::{Rebind, WatchService};
+use watch_service::WatchService;
 use futures_watch::Watch;
+
+mod client;
+use self::client::BindClient;
 
 type DestinationServiceQuery<T> = Remote<PbUpdate, T>;
 type UpdateRx<T> = Receiver<PbUpdate, T>;
-
-/// Type of the client service stack used to make destination requests.
-type ClientService = AddOrigin<Backoff<LogErrors<Reconnect<
-        tower_h2::client::Connect<
-            Timeout<LookupAddressAndConnect>,
-            ::logging::ContextualExecutor<
-                ::logging::Client<
-                    &'static str,
-                    HostAndPort
-                >
-            >,
-            BoxBody,
-        >
-    >>>>;
 
 /// Satisfies resolutions as requested via `request_rx`.
 ///
@@ -87,15 +70,6 @@ struct DestinationSet<T: HttpService<ResponseBody = RecvBody>> {
     query: Option<DestinationServiceQuery<T>>,
     dns_query: Option<IpAddrListFuture>,
     responders: Vec<Responder>,
-}
-
-/// The state needed to bind a new controller client stack.
-struct BindClient {
-    backoff_delay: Duration,
-    identity: Conditional<tls::Identity, tls::ReasonForNoTls>,
-    host_and_port: HostAndPort,
-    dns_resolver: dns::Resolver,
-    log_ctx: ::logging::Client<&'static str, HostAndPort>,
 }
 
 /// Returns a new discovery background task.
@@ -595,64 +569,6 @@ impl<T: HttpService<ResponseBody = RecvBody>> DestinationSet<T> {
             sent.is_ok()
         });
     }
-}
-
-// ===== impl BindClient =====
-
-impl BindClient {
-    fn new(
-        identity: Conditional<tls::Identity, tls::ReasonForNoTls>,
-        dns_resolver: &dns::Resolver,
-        host_and_port: HostAndPort,
-        backoff_delay: Duration,
-    ) -> Self {
-        let log_ctx = ::logging::admin().client("control", host_and_port.clone());
-        Self {
-            backoff_delay,
-            identity,
-            dns_resolver: dns_resolver.clone(),
-            host_and_port,
-            log_ctx,
-        }
-    }
-}
-impl Rebind<tls::ConditionalClientConfig> for BindClient {
-    type Service = ClientService;
-    fn rebind(
-        &mut self,
-        client_cfg: &tls::ConditionalClientConfig,
-    ) -> Self::Service {
-        let conn_cfg = match (&self.identity, client_cfg) {
-            (Conditional::Some(ref id), Conditional::Some(ref cfg)) =>
-                Conditional::Some(tls::ConnectionConfig {
-                    server_identity: id.clone(),
-                    config: cfg.clone(),
-                }),
-            (Conditional::None(ref reason), _) |
-            (_, Conditional::None(ref reason)) =>
-                Conditional::None(reason.clone()),
-        };
-        let scheme = http::uri::Scheme::from_shared(Bytes::from_static(b"http")).unwrap();
-        let authority = http::uri::Authority::from(&self.host_and_port);
-        let connect = Timeout::new(
-            LookupAddressAndConnect::new(self.host_and_port.clone(),
-                                         self.dns_resolver.clone(),
-                                         conn_cfg),
-            Duration::from_secs(3),
-        );
-        let h2_client = tower_h2::client::Connect::new(
-            connect,
-            h2::client::Builder::default(),
-            self.log_ctx.clone().executor()
-        );
-
-        let reconnect = Reconnect::new(h2_client);
-        let log_errors = LogErrors::new(reconnect);
-        let backoff = Backoff::new(log_errors, self.backoff_delay);
-        // TODO: Use AddOrigin in tower-http
-        AddOrigin::new(scheme, authority, backoff)
-    }
-
 }
 
 /// Construct a new labeled `SocketAddr `from a protobuf `WeightedAddr`.
